resolve #19, #26

+++ Models/Balance.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public double BitcoinBalance { get; set; }
-        public int USDBalance { get; set; }
-        public int DefimaBalance { get; set; }
+        public double USDBalance { get; set; }
+        public double DefimaBalance { get; set; }
         public double RateBTC_USD { get; set; }
         public double RateUSD_DEF { get; set; }
         public User User { get; set; }
